GeminiStudyAssistant/requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
pandas==2.1.1
scikit-learn==1.3.2
textstat==0.7.3
Werkzeug==2.3.7
